Fix: Rename custom validator class and package
