+++ src/main/java/ru/yandex/practicum/filmorate/validator/IsAfter.java
@@ -1,4 +1,4 @@
-package ru.yandex.practicum.filmorate.constraint;
+package ru.yandex.practicum.filmorate.validator;
 
 import jakarta.validation.Constraint;
 import jakarta.validation.Payload;
@@ -11,7 +11,7 @@ import java.lang.annotation.Target;
 
 @Target(ElementType.FIELD)
 @Retention(RetentionPolicy.RUNTIME)
-@Constraint(validatedBy = DateValidator.class)
+@Constraint(validatedBy = IsAfterValidator.class)
 @Documented
 public @interface IsAfter {
